+++ models.py
@@ -34,6 +34,6 @@ class Pet(db.Model):
     name = db.Column(db.String(30), nullable=False)
     species = db.Column(db.Text, nullable=False)
     photo_url = db.Column(db.Text, server_default=DEFAULT_IMG)
-    age = db.Column(db.Integer, CheckConstraint("age >= 0"))
+    age = db.Column(db.Float, CheckConstraint("age >= 0"))
     notes = db.Column(db.Text)
     available = db.Column(db.Boolean, nullable=False, server_default="True")
